Rename JsonType to AnyType
